+++ tests/test_io.py
@@ -12,7 +12,7 @@ from authorityspoke.holdings import Holding
 from authorityspoke.opinions import Opinion
 from authorityspoke.predicates import Predicate
 from authorityspoke.procedures import Procedure
-from authorityspoke.io.readers import get_directory_path, read_dict, read_json
+from authorityspoke.io.readers import get_directory_path, read_dict, json_holdings
 from authorityspoke.rules import Rule
 from authorityspoke.selectors import TextQuoteSelector
 
@@ -59,20 +59,20 @@ class TestEntityImport:
 
 class TestEnactmentImport:
     def test_enactment_import(self, make_regime):
-        holdings = read_json("holding_cardenas.json", regime=make_regime)
+        holdings = json_holdings("holding_cardenas.json", regime=make_regime)
         enactment_list = holdings[0].enactments
         assert "all relevant evidence is admissible" in enactment_list[0].text
 
 
 class TestFactorImport:
     def test_fact_import(self, make_regime):
-        holdings = read_json("holding_watt.json", regime=make_regime)
+        holdings = json_holdings("holding_watt.json", regime=make_regime)
         new_fact = holdings[0].inputs[1]
         assert "lived at <Hideaway Lodge>" in str(new_fact)
         assert isinstance(new_fact.context_factors[0], Entity)
 
     def test_fact_with_quantity(self, make_regime):
-        holdings = read_json("holding_watt.json", regime=make_regime)
+        holdings = json_holdings("holding_watt.json", regime=make_regime)
         new_fact = holdings[1].inputs[3]
         assert "was no more than 35 foot" in str(new_fact)
 
@@ -115,7 +115,7 @@ class TestRuleImport:
         """
 
         watt = make_opinion["watt_majority"]
-        watt.exposit(read_json("holding_watt.json"))
+        watt.exposit(json_holdings("holding_watt.json"))
         assert watt.holdings[0] == real_holding["h1"]
 
     def test_same_enactment_objects_equal(self, make_opinion_with_holding):
@@ -141,9 +141,9 @@ class TestRuleImport:
 
     def test_same_enactment_in_two_opinions(self, make_regime, make_opinion):
         watt = make_opinion["watt_majority"]
-        watt.exposit(read_json("holding_watt.json", regime=make_regime))
+        watt.exposit(json_holdings("holding_watt.json", regime=make_regime))
         brad = make_opinion["brad_majority"]
-        brad.exposit(read_json("holding_brad.json", regime=make_regime))
+        brad.exposit(json_holdings("holding_brad.json", regime=make_regime))
         assert any(
             watt.holdings[0].enactments[0].means(brad_enactment)
             for brad_enactment in brad.holdings[0].enactments
@@ -158,20 +158,20 @@ class TestRuleImport:
         """
         brad = make_opinion["brad_majority"]
         brad.holdings = []
-        brad.exposit(read_json("holding_brad.json", regime=make_regime))
+        brad.exposit(json_holdings("holding_brad.json", regime=make_regime))
         assert any(brad.holdings[6].inputs[0] == x for x in brad.holdings[5].inputs)
         assert any(brad.holdings[6].inputs[0] is x for x in brad.holdings[5].inputs)
 
     def test_use_int_not_pint_without_dimension(self, make_regime, make_opinion):
 
         brad = make_opinion["brad_majority"]
-        brad.exposit(read_json("holding_brad.json", regime=make_regime))
+        brad.exposit(json_holdings("holding_brad.json", regime=make_regime))
         assert "dimensionless" not in str(brad.holdings[6])
         assert isinstance(brad.holdings[6].inputs[0].predicate.quantity, int)
 
     def test_opinion_posits_holding(self, make_opinion, make_regime):
         brad = make_opinion["brad_majority"]
-        brad.exposit(read_json("holding_brad.json", regime=make_regime))
+        brad.exposit(json_holdings("holding_brad.json", regime=make_regime))
         assert "warrantless search and seizure" in str(brad.holdings[0])
 
     def test_opinion_posits_holding_tuple_context(
@@ -183,7 +183,7 @@ class TestRuleImport:
         """
         watt = make_opinion["watt_majority"]
         brad = make_opinion["brad_majority"]
-        brad.exposit(read_json("holding_brad.json", regime=make_regime))
+        brad.exposit(json_holdings("holding_brad.json", regime=make_regime))
         context_holding = brad.holdings[6].new_context(
             [make_entity["watt"], make_entity["trees"], make_entity["motel"]]
         )
@@ -205,7 +205,7 @@ class TestRuleImport:
         """
         watt = make_opinion["watt_majority"]
         brad = make_opinion["brad_majority"]
-        brad.exposit(read_json("holding_brad.json", regime=make_regime))
+        brad.exposit(json_holdings("holding_brad.json", regime=make_regime))
         context_holding = brad.holdings[6].new_context(
             {brad.holdings[6].generic_factors[0]: make_entity["watt"]}
         )
@@ -224,7 +224,7 @@ class TestRuleImport:
         """
         watt = make_opinion["watt_majority"]
         brad = make_opinion["brad_majority"]
-        brad.exposit(read_json("holding_brad.json", regime=make_regime))
+        brad.exposit(json_holdings("holding_brad.json", regime=make_regime))
         context_change = brad.holdings[6].new_context(
             {brad.holdings[6].generic_factors[1]: make_entity["trees_specific"]}
         )
@@ -367,5 +367,5 @@ class TestNestedFactorImport:
         testimony on the jury. Hence, admission of the testimony
         concerning appellant’s use of narcotics was improper.
         """
-        cardenas_holdings = read_json("holding_cardenas.json", regime=make_regime)
+        cardenas_holdings = json_holdings("holding_cardenas.json", regime=make_regime)
         assert len(cardenas_holdings) == 2