+++ changelog.md
@@ -14,6 +14,8 @@ dev
 - Use addition operator to add Factors as Rule inputs
 - Use addition operator to add Enactments to Rules
 - Create function to consolidate list of Enactments
+- Add Union operator for Rules
+- Move functions for loading objects from JSON and XML to new I/O modules
 
 0.1.0 (2019-06-10)
 ------------------